+++ .gitignore
@@ -1,2 +1,3 @@
 web_app/templates/graph_time_period.html
 web_app/templates/graph_feature.html
+web_app/app/__pycache__
