+++ web_app/app/predict/predict_functions.py
@@ -1,3 +1,5 @@
+
+# %%
 import pandas as pd
 import joblib
 import matplotlib.pyplot as plt
