+++ web_app/templates/base.html
@@ -10,7 +10,7 @@
     </head>
     <body>
         <nav>
-            <a href="http://localhost:8000/dataset">Explore the dataset</a>
+            <a href="http://localhost:8000/dataviz">Explore the dataset</a>
             <a href="http://localhost:8000/predict">Predictions</a>
             <a href="http://localhost:8000/about">About this project</a>
         </nav>
