+++ web_app/templates/dataset.html
@@ -10,7 +10,7 @@
 </div>
 <p><strong>On this page you can select a duration and display the associated graph, or visualize the effect of certain features on the target (national demand).</strong></p>
 <div class="form">
-    <form method="post" action="/duration">
+    <form method="post" action="/dataviz/duration">
         <fieldset>
             <div>
                 <legend>Select a time period:</legend>
@@ -38,7 +38,7 @@
         {% endblock %}
     </div>
     <p></p>
-    <form method="post" action="/feature">
+    <form method="post" action="/dataviz/feature">
         <fieldset>
             <div>
                 <legend>Select a feature:</legend>
