+++ web_app/templates/predict.html
@@ -9,7 +9,7 @@
 </div>
 <p><strong>Please, give a number of period (integer).</strong></p>
 <div class="form">
-    <form method="post" action="/predict_call">
+    <form method="post" action="/predict/predict_call">
         <fieldset>
             <div>
                 <legend>Write a number of periods:</legend>
